pdf_tools_web/backend/code to copy.py
# Production-Ready Django Backend for PDF Processing

## Folder Structure:

backend/
|-- pdf_tools/
|   |-- settings.py
|   |-- urls.py
|   |-- wsgi.py
|
|-- pdf_processing/
|   |-- views.py
|   |-- utils.py
|   |-- serializers.py
|   |-- models.py
|   |-- urls.py
|
|-- manage.py

## Install Required Libraries:
```sh
pip install django djangorestframework pdf2image python-docx reportlab pillow pypdf
```

## **Merge & Split PDFs**
### **views.py**
```python
from rest_framework.decorators import api_view
from rest_framework.response import Response
from PyPDF2 import PdfMerger, PdfReader, PdfWriter
from django.http import FileResponse
import os
from django.core.files.storage import default_storage

def save_temp_file(uploaded_file):
    file_path = default_storage.save(f'temp/{uploaded_file.name}', uploaded_file)
    return file_path

@api_view(['POST'])
def merge_pdfs(request):
    merger = PdfMerger()
    file_paths = [save_temp_file(f) for f in request.FILES.getlist('files')]
    
    for file in file_paths:
        merger.append(file)
    
    output_path = 'temp/merged.pdf'
    merger.write(output_path)
    merger.close()
    
    return FileResponse(open(output_path, 'rb'), as_attachment=True, filename='merged.pdf')

@api_view(['POST'])
def split_pdf(request):
    file = save_temp_file(request.FILES['file'])
    reader = PdfReader(file)
    output_files = []
    
    for i, page in enumerate(reader.pages):
        writer = PdfWriter()
        writer.add_page(page)
        output_file = f'temp/split_{i}.pdf'
        with open(output_file, 'wb') as f:
            writer.write(f)
        output_files.append(output_file)
    
    return FileResponse(open(output_files[0], 'rb'), as_attachment=True, filename='split_page.pdf')
```

## **PDF to Word & Word to PDF**
### **utils.py**
```python
from pdf2image import convert_from_path
from docx import Document
from reportlab.pdfgen import canvas
from django.http import FileResponse

def pdf_to_word(pdf_file):
    doc = Document()
    images = convert_from_path(pdf_file)
    for image in images:
        doc.add_paragraph(image.filename)
    output_path = 'temp/converted.docx'
    doc.save(output_path)
    return output_path

def word_to_pdf(word_file):
    doc = Document(word_file)
    pdf_path = 'temp/converted.pdf'
    c = canvas.Canvas(pdf_path)
    c.drawString(100, 750, doc.paragraphs[0].text)
    c.save()
    return pdf_path
```

### **views.py** (Add to existing views)
```python
@api_view(['POST'])
def convert_pdf_to_word(request):
    file = save_temp_file(request.FILES['file'])
    output_file = pdf_to_word(file)
    return FileResponse(open(output_file, 'rb'), as_attachment=True, filename='converted.docx')

@api_view(['POST'])
def convert_word_to_pdf(request):
    file = save_temp_file(request.FILES['file'])
    output_file = word_to_pdf(file)
    return FileResponse(open(output_file, 'rb'), as_attachment=True, filename='converted.pdf')
```

## **Image to PDF & PDF to Image**
```python
from PIL import Image

def image_to_pdf(image_file):
    img = Image.open(image_file)
    pdf_path = 'temp/converted.pdf'
    img.convert('RGB').save(pdf_path)
    return pdf_path

def pdf_to_image(pdf_file, format='JPEG'):
    images = convert_from_path(pdf_file)
    output_files = []
    for i, img in enumerate(images):
        img_path = f'temp/output_{i}.{format.lower()}'
        img.save(img_path, format)
        output_files.append(img_path)
    return output_files
```

### **views.py**
```python
@api_view(['POST'])
def convert_image_to_pdf(request):
    file = save_temp_file(request.FILES['file'])
    output_file = image_to_pdf(file)
    return FileResponse(open(output_file, 'rb'), as_attachment=True, filename='converted.pdf')

@api_view(['POST'])
def convert_pdf_to_image(request):
    file = save_temp_file(request.FILES['file'])
    output_files = pdf_to_image(file)
    return FileResponse(open(output_files[0], 'rb'), as_attachment=True, filename='converted.jpg')
```

## **URL Routing**
### **urls.py**
```python
from django.urls import path
from .views import merge_pdfs, split_pdf, convert_pdf_to_word, convert_word_to_pdf, convert_image_to_pdf, convert_pdf_to_image

urlpatterns = [
    path('merge/', merge_pdfs),
    path('split/', split_pdf),
    path('pdf-to-word/', convert_pdf_to_word),
    path('word-to-pdf/', convert_word_to_pdf),
    path('image-to-pdf/', convert_image_to_pdf),
    path('pdf-to-image/', convert_pdf_to_image),
]
```

## **Dockerfile**
```dockerfile
FROM python:3.9
WORKDIR /app
COPY . .
RUN pip install -r requirements.txt
CMD ["gunicorn", "pdf_tools.wsgi"]
```

## **docker-compose.yml**
```yaml
version: '3.8'
services:
  backend:
    build: ./backend
    ports:
      - "8000:8000"
  nginx:
    image: nginx
    ports:
      - "80:80"
```

---

This **fully functional Django backend** includes: 
✅ **Merge & Split PDFs**
✅ **PDF to Word & Word to PDF**
✅ **Image to PDF & PDF to Image**
✅ **Docker Integration for Deployment**

Let me know if you need changes or explanations! 🚀
